+++ Wrappers/Python/ccpi/viewer/utils/io.py
@@ -1,30 +1,26 @@
-from PySide2 import QtCore, QtWidgets, QtGui
-from PySide2.QtCore import QThreadPool
-from PySide2.QtWidgets import QProgressDialog, QDialog, QLabel, QComboBox, QDialogButtonBox, QFormLayout, QWidget, QVBoxLayout, QGroupBox, QLineEdit, QMessageBox
-
-import os
-import time
-import numpy
-import sys
-import shutil
-
-from functools import partial
-from os.path import isfile
-
-import vtk
-
-from ccpi.viewer.utils import Converter
-
-from ccpi.viewer.utils import cilNumpyMETAImageWriter
-
-# from ccpi.viewer.QtThreading import Worker, WorkerSignals, ErrorObserver #
-from eqt.threading import Worker
-
-from vtk.util.vtkAlgorithm import VTKPythonAlgorithmBase
-
-from ccpi.viewer.utils.conversion import cilNumpyResampleReader, cilBaseResampleReader, cilNumpyCroppedReader, cilMetaImageResampleReader, cilMetaImageCroppedReader, cilBaseCroppedReader
-
 import imghdr
+import os
+import shutil
+import sys
+import time
+from functools import partial
+
+import numpy
+import vtk
+from ccpi.viewer.utils import Converter
+from ccpi.viewer.utils.conversion import (cilBaseCroppedReader,
+                                          cilBaseResampleReader,
+                                          cilMetaImageCroppedReader,
+                                          cilMetaImageResampleReader,
+                                          cilNumpyCroppedReader,
+                                          cilNumpyResampleReader)
+from eqt.threading import Worker
+from PySide2 import QtCore, QtGui
+from PySide2.QtCore import QThreadPool
+from PySide2.QtWidgets import (QComboBox, QDialog, QDialogButtonBox,
+                               QFormLayout, QGroupBox, QLabel, QLineEdit,
+                               QMessageBox, QProgressDialog, QVBoxLayout,
+                               QWidget)
 
 # ImageCreator class